requirements.txt
opencv-python
numpy
mysql-connector-python
Pillow
streamlit
shapely
ultralytics
openai
